Respond to feedback from peer review
