+++ src/Components/WebAssembly/WebAssembly/src/Services/WebAssemblyConsoleLogger.cs
@@ -9,43 +9,6 @@ using Microsoft.JSInterop;
 
 namespace Microsoft.AspNetCore.Components.WebAssembly.Services
 {
-    /// <summary>
-    /// A provider of <see cref="WebAssemblyConsoleLogger{T}"/> instances.
-    /// </summary>
-    public class WebAssemblyConsoleLoggerProvider : ILoggerProvider
-    {
-        private ConcurrentDictionary<string, WebAssemblyConsoleLogger<object>> _loggers;
-        private IJSInProcessRuntime _jsRuntime;
-        private bool _disposed;
-
-        /// <summary>
-        /// Creates an instance of <see cref="WebAssemblyConsoleLoggerProvider"/>.
-        /// </summary>
-        /// <param name="options">The options to create <see cref="WebAssemblyConsoleLogger"/> instances with.</param>
-        public WebAssemblyConsoleLoggerProvider(IJSInProcessRuntime jsRuntime)
-        {
-            _loggers = new ConcurrentDictionary<string, WebAssemblyConsoleLogger<object>>();
-            _jsRuntime = jsRuntime;
-        }
-
-        /// <inheritdoc />
-        public ILogger CreateLogger(string name)
-        {
-            return _loggers.GetOrAdd(name, loggerName => new WebAssemblyConsoleLogger<object>(name, _jsRuntime));
-        }
-
-        /// <inheritdoc />
-        public void Dispose()
-        {
-            if (!_disposed)
-            {
-                _loggers = null;
-                _jsRuntime = null;
-            }
-            _disposed = true;
-        }
-    }
-
     internal class WebAssemblyConsoleLogger<T> : ILogger<T>, ILogger
     {
         private static readonly string _loglevelPadding = ": ";
@@ -81,7 +44,7 @@ namespace Microsoft.AspNetCore.Components.WebAssembly.Services
 
         public bool IsEnabled(LogLevel logLevel)
         {
-            return logLevel >= LogLevel.Warning && logLevel != LogLevel.None;
+            return logLevel != LogLevel.None;
         }
 
         public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
